requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector==2.2.9
numpy==1.19.5
pandas==1.3.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
